requirements.txt
langgraph>=0.2.0
langchain>=0.3.0
langchain-community>=0.3.0
langchain-ollama>=0.2.0
langchain-groq>=0.2.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
chromadb>=0.5.0
sentence-transformers>=3.0.0
redis>=5.0.0
asyncpg>=0.30.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
httpx>=0.27.0
websockets>=13.0
pytest>=8.0.0
pytest-asyncio>=0.24.0
pyyaml>=6.0.0
numpy>=1.26.0
